Move Annotation helper to main code, return pair of Field,Annotation.
